+++ examples/vision-camera/config/appConfig.ts
@@ -1,0 +1,5 @@
+// Visit https://fishjam.io/app/sandbox for your app ID
+export const APP_ID = "";
+
+export const DEFAULT_ROOM_NAME = "Room";
+export const DEFAULT_PEER_NAME = "User";
